web/requirements.txt
Django==2.1.5
psycopg2==2.7.7
pytz==2018.9
django-truncate==0.1
djangorestframework==3.9.1